+++ clang/test/OpenMP/target_teams_distribute_num_teams_messages.cpp
@@ -44,6 +44,9 @@ T tmain(T argc) {
 #pragma omp target teams distribute num_teams(3.14) // expected-error 2 {{expression must have integral or unscoped enumeration type, not 'double'}}
   for (int i=0; i<100; i++) foo();
 
+#pragma omp target teams distribute num_teams(1, 2, 3) // expected-error {{only one expression allowed to 'num_teams' clause}}
+  for (int i=0; i<100; i++) foo();
+
   return 0;
 }
 
@@ -83,6 +86,9 @@ int main(int argc, char **argv) {
   for (int i=0; i<100; i++) foo();
 
 #pragma omp target teams distribute num_teams (3.14) // expected-error {{expression must have integral or unscoped enumeration type, not 'double'}}
+  for (int i=0; i<100; i++) foo();
+
+#pragma omp target teams distribute num_teams(1, 2, 3) // expected-error {{only one expression allowed to 'num_teams' clause}}
   for (int i=0; i<100; i++) foo();
 
   return tmain<int, 10>(argc); // expected-note {{in instantiation of function template specialization 'tmain<int, 10>' requested here}}